requirements.txt
sqlalchemy
sqlalchemy_utils
python-dotenv
pydantic
fastapi
